prediction_cote_final/requirements.txt
# packaging
pip>=9
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn

#scrapping
request
bs4


#############
time
datetime
os
argparse